fixed id issue

+++ src/main/java/com/example/demo/bean/Person.java
@@ -17,7 +17,7 @@ public class Person {
 
     @Id
     @GeneratedValue
-    private int id;
+    private Long id;
     private String name;
     private String location;
     private Date birthDate;
@@ -28,11 +28,11 @@ public class Person {
     @CreationTimestamp
     private LocalDateTime createdDate;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -93,7 +93,7 @@ public class Person {
         this.createdDate = createdDate;
     }
 
-    public Person(int id, String name, String location, Date birthDate) {
+    public Person(Long id, String name, String location, Date birthDate) {
         this.id = id;
         this.name = name;
         this.location = location;
